+++ NEWS.md
@@ -1,7 +1,8 @@
-# tidytext (development version)
+# tidytext (development version, to be released as 0.4.0)
 
 * hunspell is now a suggested dependency, thanks to @MichaelChirico (#221)
 * Added `stm()` tidiers for high FREX and lift words (#223)
+* Removed tweet-specific tokenizers because of changes in upstream dependencies
 
 # tidytext 0.3.4